pyproject.toml
[project]
name = "multi-restaurant-alx-captsone"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "black>=25.12.0",
    "django>=6.0",
    "django-silk>=5.4.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
]
